deploy_package/requirements.txt
# Temu Deals Bot - Requirements
# Free hosting compatible
python-telegram-bot==20.3
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
python-dotenv==1.0.0
apscheduler==3.10.4
